sdh commutation: containers VC4 and VC3

+++ ManagementApp/NetNode/FIB.cs
@@ -10,10 +10,10 @@ namespace NetNode
     //forwarding information table: destination mask oport
     class FIB
     {
-        public double in_cont;
+        public int in_cont;
         public int oport;
-        public double out_cont;
-        public FIB(double in_cont, int oport, double out_cont)
+        public int out_cont;
+        public FIB(int in_cont, int oport, int out_cont)
         {
             this.in_cont = in_cont;
             this.oport = oport;
